Rename src/tests/test.h -> src/tests/app_test.h to avoid confusion with observable tests/test.h
