+++ src/app/cli/cli_tests.cpp
@@ -5,7 +5,7 @@
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
 
-#include "tests/test.h"
+#include "tests/app_test.h"
 
 #include "app/cli/app_options.h"
 #include "app/cli/cli_processor.h"
